feat(Importer): create file to import the data to trading and proceed
